working on tests
