+++ tests/sentry/incidents/endpoints/test_organization_alert_rule_details.py
@@ -1,8 +1,13 @@
+from __future__ import annotations
+
 from copy import deepcopy
 from functools import cached_property
+from typing import Any, Optional
+from unittest.mock import patch
 
 import responses
 from django.conf import settings
+from rest_framework.exceptions import ErrorDetail
 
 from sentry import audit_log
 from sentry.api.serializers import serialize
@@ -16,17 +21,61 @@ from sentry.incidents.models import (
     IncidentStatus,
 )
 from sentry.incidents.serializers import AlertRuleSerializer
-from sentry.models import AuditLogEntry
+from sentry.integrations.slack.client import SlackClient
+from sentry.models import AuditLogEntry, Integration
+from sentry.models.actor import get_actor_for_user
 from sentry.services.hybrid_cloud.app import app_service
+from sentry.shared_integrations.exceptions.base import ApiError
 from sentry.silo import SiloMode
 from sentry.testutils.cases import APITestCase
 from sentry.testutils.outbox import outbox_runner
 from sentry.testutils.silo import assume_test_silo_mode, region_silo_test
+from sentry.utils import json
 from tests.sentry.incidents.endpoints.test_organization_alert_rule_index import AlertRuleBase
 
 
 class AlertRuleDetailsBase(AlertRuleBase):
     endpoint = "sentry-api-0-organization-alert-rule-details"
+
+    def setUp(self):
+        self.owner_user = self.create_user(email="ckjfg@dkfjghsg.com")
+        self.create_member(
+            user=self.owner_user, organization=self.organization, role="owner", teams=[self.team]
+        )
+        # Default to the 'owner' user
+        self.user = self.owner_user
+        self.valid_params = {
+            "name": "hello",
+            "time_window": 10,
+            "query": "level:error",
+            "threshold_type": 0,
+            "resolve_threshold": 100,
+            "alert_threshold": 0,
+            "aggregate": "count_unique(user)",
+            "threshold_period": 1,
+            "projects": [self.project.slug],
+            "triggers": [
+                {
+                    "label": "critical",
+                    "alertThreshold": 200,
+                    "actions": [
+                        {"type": "email", "targetType": "team", "targetIdentifier": self.team.id}
+                    ],
+                },
+                {
+                    "label": "warning",
+                    "alertThreshold": 150,
+                    "actions": [
+                        {"type": "email", "targetType": "team", "targetIdentifier": self.team.id},
+                        {
+                            "type": "email",
+                            "targetType": "user",
+                            "targetIdentifier": self.owner_user.id,
+                        },
+                    ],
+                },
+            ],
+        }
 
     def new_alert_rule(self, data=None):
         if data is None:
@@ -71,10 +120,7 @@ class AlertRuleDetailsBase(AlertRuleBase):
         return self.new_alert_rule(data=deepcopy(self.alert_rule_dict))
 
     def test_invalid_rule_id(self):
-        self.create_member(
-            user=self.user, organization=self.organization, role="owner", teams=[self.team]
-        )
-        self.login_as(self.user)
+        self.login_as(self.owner_user)
         with self.feature("organizations:incidents"):
             resp = self.get_response(self.organization.slug, 1234)
 
@@ -89,10 +135,7 @@ class AlertRuleDetailsBase(AlertRuleBase):
         assert resp.status_code == 403
 
     def test_no_feature(self):
-        self.create_member(
-            user=self.user, organization=self.organization, role="owner", teams=[self.team]
-        )
-        self.login_as(self.user)
+        self.login_as(self.owner_user)
         resp = self.get_response(self.organization.slug, self.alert_rule.id)
         assert resp.status_code == 404
 
@@ -106,6 +149,14 @@ class AlertRuleDetailsGetEndpointTest(AlertRuleDetailsBase, APITestCase):
             resp = self.get_success_response(self.organization.slug, self.alert_rule.id)
 
         assert resp.data == serialize(self.alert_rule, serializer=DetailedAlertRuleSerializer())
+
+    def test_aggregate_translation(self):
+        self.login_as(self.user)
+        alert_rule = self.create_alert_rule(aggregate="count_unique(tags[sentry:user])")
+        with self.feature("organizations:incidents"):
+            resp = self.get_success_response(self.organization.slug, alert_rule.id)
+            assert resp.data["aggregate"] == "count_unique(user)"
+            assert alert_rule.snuba_query.aggregate == "count_unique(tags[sentry:user])"
 
     def test_expand_latest_incident(self):
         self.create_team(organization=self.organization, members=[self.user])
@@ -126,6 +177,116 @@ class AlertRuleDetailsGetEndpointTest(AlertRuleDetailsBase, APITestCase):
         assert resp.data["latestIncident"] is not None
         assert resp.data["latestIncident"]["id"] == str(incident.id)
         assert "latestIncident" not in no_expand_resp.data
+
+    def test_with_snooze_rule(self):
+        self.create_team(organization=self.organization, members=[self.user])
+        self.login_as(self.user)
+        self.snooze_rule(user_id=self.user.id, owner_id=self.user.id, alert_rule=self.alert_rule)
+
+        with self.feature("organizations:incidents"):
+            response = self.get_success_response(self.organization.slug, self.alert_rule.id)
+
+        assert response.data["snooze"]
+        assert response.data["snoozeCreatedBy"] == "You"
+
+    def test_with_snooze_rule_everyone(self):
+        self.create_team(organization=self.organization, members=[self.user])
+        self.login_as(self.user)
+
+        user2 = self.create_user("user2@example.com")
+        self.snooze_rule(owner_id=user2.id, alert_rule=self.alert_rule)
+
+        with self.feature("organizations:incidents"):
+            response = self.get_success_response(self.organization.slug, self.alert_rule.id)
+
+        assert response.data["snooze"]
+        assert response.data["snoozeCreatedBy"] == user2.get_display_name()
+
+
+@region_silo_test(stable=True)
+class AlertRuleDetailsPutEndpointTest(AlertRuleDetailsBase, APITestCase):
+    method = "put"
+
+    def test_simple(self):
+        self.login_as(self.owner_user)
+        alert_rule = self.alert_rule
+        # We need the IDs to force update instead of create, so we just get the rule using our own API. Like frontend would.
+        serialized_alert_rule = self.get_serialized_alert_rule()
+        serialized_alert_rule["name"] = "what"
+
+        with self.feature("organizations:incidents"), outbox_runner():
+            resp = self.get_success_response(
+                self.organization.slug, alert_rule.id, **serialized_alert_rule
+            )
+
+        alert_rule.name = "what"
+        alert_rule.date_modified = resp.data["dateModified"]
+        assert resp.data == serialize(alert_rule)
+        assert resp.data["name"] == "what"
+        assert resp.data["dateModified"] > serialized_alert_rule["dateModified"]
+
+        with assume_test_silo_mode(SiloMode.CONTROL):
+            audit_log_entry = AuditLogEntry.objects.filter(
+                event=audit_log.get_event_id("ALERT_RULE_EDIT"), target_object=alert_rule.id
+            )
+        assert len(audit_log_entry) == 1
+        assert (
+            resp.renderer_context["request"].META["REMOTE_ADDR"]
+            == list(audit_log_entry)[0].ip_address
+        )
+
+    def test_change_name_of_existing_alert(self):
+        self.login_as(self.owner_user)
+        test_params = self.valid_params.copy()
+        test_params["resolve_threshold"] = self.alert_rule.resolve_threshold
+        test_params.update({"name": "what"})
+
+        with self.feature("organizations:incidents"), outbox_runner():
+            resp = self.get_success_response(
+                self.organization.slug, self.alert_rule.id, **test_params
+            )
+
+        self.alert_rule.refresh_from_db()
+        self.alert_rule.name = "what"
+        self.alert_rule.snuba_query.refresh_from_db()
+        assert resp.data == serialize(self.alert_rule)
+        assert resp.data["name"] == "what"
+
+        # We validate that there's only been one change to the alert
+        with assume_test_silo_mode(SiloMode.CONTROL):
+            audit_log_entry = AuditLogEntry.objects.filter(
+                event=audit_log.get_event_id("ALERT_RULE_EDIT"), target_object=resp.data["id"]
+            )
+        assert len(audit_log_entry) == 1
+
+    def test_sentry_app(self):
+        sentry_app = self.create_sentry_app(
+            name="foo", organization=self.organization, is_alertable=True, verify_install=False
+        )
+        self.create_sentry_app_installation(
+            slug=sentry_app.slug, organization=self.organization, user=self.user
+        )
+        self.login_as(self.owner_user)
+        alert_rule = self.alert_rule
+        # We need the IDs to force update instead of create, so we just get the rule using our own API. Like frontend would.
+        serialized_alert_rule = self.get_serialized_alert_rule()
+        serialized_alert_rule["name"] = "ValidSentryAppTestRule"
+        serialized_alert_rule["triggers"][0]["actions"][0] = {
+            "type": "sentry_app",
+            "targetType": "sentry_app",
+            "targetIdentifier": sentry_app.id,
+            "sentryAppId": sentry_app.id,
+        }
+
+        with self.feature("organizations:incidents"):
+            resp = self.get_success_response(
+                self.organization.slug, alert_rule.id, **serialized_alert_rule
+            )
+
+        alert_rule.refresh_from_db()
+        alert_rule.name = "ValidSentryAppTestRule"
+        assert resp.data == serialize(alert_rule)
+        assert resp.data["triggers"][0]["actions"][0]["sentryAppId"] == sentry_app.id
 
     @responses.activate
     def test_with_unresponsive_sentryapp(self):
@@ -179,105 +340,39 @@ class AlertRuleDetailsGetEndpointTest(AlertRuleDetailsBase, APITestCase):
         )
         assert resp.data["triggers"][0]["actions"][0]["disabled"] is True
 
-    def test_with_snooze_rule(self):
-        self.create_team(organization=self.organization, members=[self.user])
-        self.login_as(self.user)
-        self.snooze_rule(user_id=self.user.id, owner_id=self.user.id, alert_rule=self.alert_rule)
-
-        with self.feature("organizations:incidents"):
-            response = self.get_success_response(self.organization.slug, self.alert_rule.id)
-
-        assert response.data["snooze"]
-        assert response.data["snoozeCreatedBy"] == "You"
-
-    def test_with_snooze_rule_everyone(self):
-        self.create_team(organization=self.organization, members=[self.user])
-        self.login_as(self.user)
-
-        user2 = self.create_user("user2@example.com")
-        self.snooze_rule(owner_id=user2.id, alert_rule=self.alert_rule)
-
-        with self.feature("organizations:incidents"):
-            response = self.get_success_response(self.organization.slug, self.alert_rule.id)
-
-        assert response.data["snooze"]
-        assert response.data["snoozeCreatedBy"] == user2.get_display_name()
-
-
-@region_silo_test(stable=True)
-class AlertRuleDetailsPutEndpointTest(AlertRuleDetailsBase, APITestCase):
-    method = "put"
-
-    def test_simple(self):
-        self.create_member(
-            user=self.user, organization=self.organization, role="owner", teams=[self.team]
-        )
-
-        self.login_as(self.user)
-        alert_rule = self.alert_rule
-        # We need the IDs to force update instead of create, so we just get the rule using our own API. Like frontend would.
-        serialized_alert_rule = self.get_serialized_alert_rule()
-        serialized_alert_rule["name"] = "what"
-
-        with self.feature("organizations:incidents"), outbox_runner():
-            resp = self.get_success_response(
-                self.organization.slug, alert_rule.id, **serialized_alert_rule
-            )
-
-        alert_rule.name = "what"
-        alert_rule.date_modified = resp.data["dateModified"]
-        assert resp.data == serialize(alert_rule)
-        assert resp.data["name"] == "what"
-        assert resp.data["dateModified"] > serialized_alert_rule["dateModified"]
-
-        with assume_test_silo_mode(SiloMode.CONTROL):
-            audit_log_entry = AuditLogEntry.objects.filter(
-                event=audit_log.get_event_id("ALERT_RULE_EDIT"), target_object=alert_rule.id
-            )
-        assert len(audit_log_entry) == 1
-        assert (
-            resp.renderer_context["request"].META["REMOTE_ADDR"]
-            == list(audit_log_entry)[0].ip_address
-        )
-
-    def test_sentry_app(self):
-        self.create_member(
-            user=self.user, organization=self.organization, role="owner", teams=[self.team]
-        )
-        sentry_app = self.create_sentry_app(
-            name="foo", organization=self.organization, is_alertable=True, verify_install=False
-        )
+    def test_no_config_sentry_app(self):
+        self.login_as(self.owner_user)
+        sentry_app = self.create_sentry_app(is_alertable=True)
         self.create_sentry_app_installation(
             slug=sentry_app.slug, organization=self.organization, user=self.user
         )
-        self.login_as(self.user)
-        alert_rule = self.alert_rule
-        # We need the IDs to force update instead of create, so we just get the rule using our own API. Like frontend would.
-        serialized_alert_rule = self.get_serialized_alert_rule()
-        serialized_alert_rule["name"] = "ValidSentryAppTestRule"
-        serialized_alert_rule["triggers"][0]["actions"][0] = {
-            "type": "sentry_app",
-            "targetType": "sentry_app",
-            "targetIdentifier": sentry_app.id,
-            "sentryAppId": sentry_app.id,
+        test_params = {
+            **self.valid_params,
+            "triggers": [
+                {
+                    "actions": [
+                        {
+                            "type": "sentry_app",
+                            "targetType": "sentry_app",
+                            "targetIdentifier": sentry_app.id,
+                            "sentryAppId": sentry_app.id,
+                        }
+                    ],
+                    "alertThreshold": 300,
+                    "label": "critical",
+                }
+            ],
         }
-
-        with self.feature("organizations:incidents"):
-            resp = self.get_success_response(
-                self.organization.slug, alert_rule.id, **serialized_alert_rule
+        with self.feature(["organizations:incidents", "organizations:performance-view"]):
+            self.get_success_response(
+                self.organization.slug,
+                self.alert_rule.id,
+                status_code=200,
+                **test_params,
             )
 
-        alert_rule.refresh_from_db()
-        alert_rule.name = "ValidSentryAppTestRule"
-        assert resp.data == serialize(alert_rule)
-        assert resp.data["triggers"][0]["actions"][0]["sentryAppId"] == sentry_app.id
-
     def test_not_updated_fields(self):
-        self.create_member(
-            user=self.user, organization=self.organization, role="owner", teams=[self.team]
-        )
-
-        self.login_as(self.user)
+        self.login_as(self.owner_user)
         alert_rule = self.alert_rule
         # We need the IDs to force update instead of create, so we just get the rule using our own API. Like frontend would.
         serialized_alert_rule = self.get_serialized_alert_rule()
@@ -298,11 +393,7 @@ class AlertRuleDetailsPutEndpointTest(AlertRuleDetailsBase, APITestCase):
         assert updated_sub.subscription_id == existing_sub.subscription_id
 
     def test_update_trigger_label_to_unallowed_value(self):
-        self.create_member(
-            user=self.user, organization=self.organization, role="owner", teams=[self.team]
-        )
-
-        self.login_as(self.user)
+        self.login_as(self.owner_user)
 
         alert_rule = self.alert_rule
         # We need the IDs to force update instead of create, so we just get the rule using our own API. Like frontend would.
@@ -322,11 +413,7 @@ class AlertRuleDetailsPutEndpointTest(AlertRuleDetailsBase, APITestCase):
             assert resp.data == {"nonFieldErrors": ['Trigger 2 must be labeled "warning"']}
 
     def test_update_trigger_alert_threshold(self):
-        self.create_member(
-            user=self.user, organization=self.organization, role="owner", teams=[self.team]
-        )
-
-        self.login_as(self.user)
+        self.login_as(self.owner_user)
         alert_rule = self.alert_rule
         # We need the IDs to force update instead of create, so we just get the rule using our own API. Like frontend would.
         serialized_alert_rule = self.get_serialized_alert_rule()
@@ -344,11 +431,7 @@ class AlertRuleDetailsPutEndpointTest(AlertRuleDetailsBase, APITestCase):
 
     def test_delete_resolve_alert_threshold(self):
         # This is a test to make sure we can remove a resolveThreshold after it has been set.
-        self.create_member(
-            user=self.user, organization=self.organization, role="owner", teams=[self.team]
-        )
-
-        self.login_as(self.user)
+        self.login_as(self.owner_user)
         alert_rule = self.alert_rule
         alert_rule.update(resolve_threshold=75)
         # We need the IDs to force update instead of create, so we just get the rule using our own API. Like frontend would.
@@ -367,11 +450,7 @@ class AlertRuleDetailsPutEndpointTest(AlertRuleDetailsBase, APITestCase):
 
     def test_update_resolve_alert_threshold(self):
         # This is a test to make sure we can remove a resolveThreshold after it has been set.
-        self.create_member(
-            user=self.user, organization=self.organization, role="owner", teams=[self.team]
-        )
-
-        self.login_as(self.user)
+        self.login_as(self.owner_user)
         alert_rule = self.alert_rule
         alert_rule.update(resolve_threshold=75)
         # We need the IDs to force update instead of create, so we just get the rule using our own API. Like frontend would.
@@ -388,11 +467,7 @@ class AlertRuleDetailsPutEndpointTest(AlertRuleDetailsBase, APITestCase):
         assert resp.data["resolveThreshold"] == 75
 
     def test_delete_trigger(self):
-        self.create_member(
-            user=self.user, organization=self.organization, role="owner", teams=[self.team]
-        )
-
-        self.login_as(self.user)
+        self.login_as(self.owner_user)
         alert_rule = self.alert_rule
         # We need the IDs to force update instead of create, so we just get the rule using our own API. Like frontend would.
         serialized_alert_rule = self.get_serialized_alert_rule()
@@ -407,11 +482,7 @@ class AlertRuleDetailsPutEndpointTest(AlertRuleDetailsBase, APITestCase):
         assert len(resp.data["triggers"]) == 1
 
     def test_delete_action(self):
-        self.create_member(
-            user=self.user, organization=self.organization, role="owner", teams=[self.team]
-        )
-
-        self.login_as(self.user)
+        self.login_as(self.owner_user)
         alert_rule = self.alert_rule
         # We need the IDs to force update instead of create, so we just get the rule using our own API. Like frontend would.
         serialized_alert_rule = self.get_serialized_alert_rule()
@@ -436,11 +507,7 @@ class AlertRuleDetailsPutEndpointTest(AlertRuleDetailsBase, APITestCase):
         assert len(resp.data["triggers"][1]["actions"]) == 0
 
     def test_update_trigger_action_type(self):
-        self.create_member(
-            user=self.user, organization=self.organization, role="owner", teams=[self.team]
-        )
-
-        self.login_as(self.user)
+        self.login_as(self.owner_user)
 
         alert_rule = self.alert_rule
         # We need the IDs to force update instead of create, so we just get the rule using our own API. Like frontend would.
@@ -466,10 +533,7 @@ class AlertRuleDetailsPutEndpointTest(AlertRuleDetailsBase, APITestCase):
         assert len(resp.data["triggers"][1]["actions"]) == 2
 
     def test_invalid_thresholds(self):
-        self.create_member(
-            user=self.user, organization=self.organization, role="owner", teams=[self.team]
-        )
-        self.login_as(self.user)
+        self.login_as(self.owner_user)
         alert_rule = self.alert_rule
         # We need the IDs to force update instead of create, so we just get the rule using our own API. Like frontend would.
         serialized_alert_rule = self.get_serialized_alert_rule()
@@ -482,10 +546,7 @@ class AlertRuleDetailsPutEndpointTest(AlertRuleDetailsBase, APITestCase):
             )
 
     def test_update_snapshot(self):
-        self.create_member(
-            user=self.user, organization=self.organization, role="owner", teams=[self.team]
-        )
-        self.login_as(self.user)
+        self.login_as(self.owner_user)
         alert_rule = self.alert_rule
         # We need the IDs to force update instead of create, so we just get the rule using our own API. Like frontend would.
         serialized_alert_rule = self.get_serialized_alert_rule()
@@ -500,11 +561,7 @@ class AlertRuleDetailsPutEndpointTest(AlertRuleDetailsBase, APITestCase):
             )
 
     def test_no_owner(self):
-        self.create_member(
-            user=self.user, organization=self.organization, role="owner", teams=[self.team]
-        )
-
-        self.login_as(self.user)
+        self.login_as(self.owner_user)
         alert_rule = self.alert_rule
         # We need the IDs to force update instead of create, so we just get the rule using our own API. Like frontend would.
         serialized_alert_rule = self.get_serialized_alert_rule()
@@ -519,16 +576,343 @@ class AlertRuleDetailsPutEndpointTest(AlertRuleDetailsBase, APITestCase):
         assert resp.data == serialize(alert_rule, self.user)
         assert resp.data["owner"] is None
 
+    def test_not_updated_fields(self):
+        self.login_as(self.owner_user)
+        test_params = self.valid_params.copy()
+        test_params["resolve_threshold"] = self.alert_rule.resolve_threshold
+        test_params["aggregate"] = self.alert_rule.snuba_query.aggregate
+
+        with self.feature("organizations:incidents"):
+            resp = self.get_success_response(
+                self.organization.slug, self.alert_rule.id, **test_params
+            )
+
+        existing_sub = self.alert_rule.snuba_query.subscriptions.first()
+
+        # Alert rule should be exactly the same
+        self.alert_rule.refresh_from_db()
+        assert resp.data == serialize(self.alert_rule)
+        # If the aggregate changed we'd have a new subscription, validate that
+        # it hasn't changed explicitly
+        updated_sub = AlertRule.objects.get(id=self.alert_rule.id).snuba_query.subscriptions.first()
+        assert updated_sub.subscription_id == existing_sub.subscription_id
+
+    def _mock_slack_response(self, url: str, body: dict[str, Any], status: int = 200) -> None:
+        responses.add(
+            method=responses.GET,
+            url=url,
+            status=status,
+            content_type="application/json",
+            body=json.dumps(body),
+        )
+
+    def _organization_alert_rule_api_call(
+        self,
+        channelName: Optional[str] = None,
+        channelID: Optional[str] = None,
+    ) -> Request:
+        """
+        Call the org alert rule API but do some Slack integration set up before doing so
+        """
+        # Set up the Slack integration
+        self.integration = self.create_slack_integration(
+            self.organization,
+            external_id="TXXXXXXX1",
+            user=self.user,
+        )
+
+        # Prep steps for the API call
+        test_params = self.valid_params.copy()
+        test_params["triggers"] = [
+            {
+                "label": "critical",
+                "alertThreshold": 200,
+                "actions": [
+                    {"type": "slack", "targetType": "specific", "integration": self.integration.id}
+                ],
+            },
+        ]
+        if channelName:
+            test_params["triggers"][0]["actions"][0]["targetIdentifier"] = channelName
+        if channelID:
+            # The trigger code would accept channelId to be a string and that is why I don't cast it to an int
+            test_params["triggers"][0]["actions"][0]["inputChannelId"] = channelID
+
+        with self.feature("organizations:incidents"):
+            resp = self.get_response(self.organization.slug, self.alert_rule.id, **test_params)
+        return resp
+
+    @patch(
+        "sentry.integrations.slack.utils.channel.get_channel_id_with_timeout",
+        return_value=("#", None, True),
+    )
+    @patch("sentry.tasks.integrations.slack.find_channel_id_for_alert_rule.apply_async")
+    @patch("sentry.integrations.slack.utils.rule_status.uuid4")
+    def test_kicks_off_slack_async_job(
+        self, mock_uuid4, mock_find_channel_id_for_alert_rule, mock_get_channel_id
+    ):
+        mock_uuid4.return_value = self.get_mock_uuid()
+        self.integration = Integration.objects.create(
+            provider="slack",
+            name="Team A",
+            external_id="TXXXXXXX1",
+            metadata={"access_token": "xoxp-xxxxxxxxx-xxxxxxxxxx-xxxxxxxxxxxx"},
+        )
+        self.integration.add_organization(self.organization, self.user)
+        test_params = self.valid_params.copy()
+        test_params["triggers"] = [
+            {
+                "label": "critical",
+                "alertThreshold": 200,
+                "actions": [
+                    {
+                        "type": "slack",
+                        "targetIdentifier": "my-channel",
+                        "targetType": "specific",
+                        "integration": self.integration.id,
+                    }
+                ],
+            },
+        ]
+
+        with self.feature("organizations:incidents"):
+            resp = self.get_response(self.organization.slug, self.alert_rule.id, **test_params)
+
+        # A task with this uuid has been scheduled because there's a Slack channel async search
+        assert resp.data["uuid"] == "abc123"
+        kwargs = {
+            "organization_id": self.organization.id,
+            "uuid": "abc123",
+            "alert_rule_id": self.alert_rule.id,
+            "data": test_params,
+            "user_id": self.owner_user.id,
+        }
+        mock_find_channel_id_for_alert_rule.assert_called_once_with(kwargs=kwargs)
+
+    @responses.activate
+    def test_create_slack_alert_with_name_and_channel_id(self):
+        """
+        The user specifies the Slack channel and channel ID (which match).
+        """
+        channelName = "my-channel"
+        # Specifying an inputChannelID will cause the validate_channel_id logic to be triggered
+        channelID = 123
+        self._mock_slack_response(
+            url=f"https://slack.com/api/conversations.info?channel={channelID}",
+            body={"ok": "true", "channel": {"name": channelName}},
+        )
+        resp = self._organization_alert_rule_api_call(channelName, channelID)
+
+        stored_action = resp.data["triggers"][0]["actions"][0]
+        assert stored_action["inputChannelId"] == str(channelID)
+        assert stored_action["targetIdentifier"] == channelName
+
+    @responses.activate
+    def test_create_slack_alert_with_mismatch_name_and_channel_id(self):
+        """
+        The user specifies the Slack channel and channel ID but they do not match.
+        """
+        otherChannel = "some-other-channel"
+        channelName = "my-channel"
+        # Specifying an inputChannelID will cause the validate_channel_id logic to be triggered
+        channelID = 123
+        self._mock_slack_response(
+            url=f"https://slack.com/api/conversations.info?channel={channelID}",
+            body={"ok": "true", "channel": {"name": otherChannel}},
+        )
+        resp = self._organization_alert_rule_api_call(channelName, channelID)
+
+        assert resp.status_code == 400
+        assert resp.data == {
+            "nonFieldErrors": [
+                ErrorDetail(
+                    string=f"Received channel name {otherChannel} does not match inputted channel name {channelName}.",
+                    code="invalid",
+                )
+            ]
+        }
+
+    # An incorrect channelID will raise an ApiError in the Slack client
+    @patch.object(SlackClient, "get", side_effect=ApiError(text="channel_not_found"))
+    @responses.activate
+    def test_create_slack_alert_with_non_existent_channel_id(self, mock_slack_client):
+        """
+        The user specifies a bad Slack channel ID.
+        """
+        channelName = "my-channel"
+        # Specifying an inputChannelID will cause the validate_channel_id logic to be triggered
+        channelID = 123
+        resp = self._organization_alert_rule_api_call(channelName, channelID)
+
+        assert resp.status_code == 400
+        assert resp.data == {
+            "nonFieldErrors": [
+                ErrorDetail(string="Channel not found. Invalid ID provided.", code="invalid")
+            ]
+        }
+
+    @patch("sentry.tasks.integrations.slack.find_channel_id_for_alert_rule.apply_async")
+    @patch("sentry.integrations.slack.utils.rule_status.uuid4")
+    @responses.activate
+    def test_create_slack_alert_with_empty_channel_id(
+        self, mock_uuid4, mock_find_channel_id_for_alert_rule
+    ):
+        """
+        The user selects the channel ID field and the UI will send the empty string to the
+        endpoint, thus, a channel name search will be performed
+        """
+        mock_uuid4.return_value = self.get_mock_uuid()
+        channelName = "my-channel"
+        # Because channel ID is an empty string it will be converted to an async request for the channel name
+        resp = self._organization_alert_rule_api_call(channelName, channelID="")
+
+        # A task with this uuid has been scheduled because there's a Slack channel async search
+        assert resp.status_code == 202
+        assert resp.data == {"uuid": "abc123"}
+
+    @patch(
+        "sentry.integrations.slack.utils.channel.get_channel_id_with_timeout",
+        side_effect=[("#", 10, False), ("#", 10, False), ("#", 20, False)],
+    )
+    @patch("sentry.integrations.slack.utils.rule_status.uuid4")
+    def test_async_lookup_outside_transaction(self, mock_uuid4, mock_get_channel_id):
+        mock_uuid4.return_value = self.get_mock_uuid()
+        with assume_test_silo_mode(SiloMode.CONTROL):
+            self.integration = Integration.objects.create(
+                provider="slack",
+                name="Team A",
+                external_id="TXXXXXXX1",
+                metadata={"access_token": "xoxp-xxxxxxxxx-xxxxxxxxxx-xxxxxxxxxxxx"},
+            )
+            self.integration.add_organization(self.organization, self.user)
+        test_params = self.valid_params.copy()
+        test_params["triggers"] = [
+            {
+                "label": "critical",
+                "alertThreshold": 200,
+                "actions": [
+                    {
+                        "type": "slack",
+                        "targetIdentifier": "my-channel",
+                        "targetType": "specific",
+                        "integration": self.integration.id,
+                    },
+                ],
+            },
+        ]
+
+        with self.feature("organizations:incidents"), self.tasks():
+            resp = self.get_response(self.organization.slug, self.alert_rule.id, **test_params)
+
+        # A task with this uuid has been scheduled because there's a Slack channel async search
+        assert resp.data["uuid"] == "abc123"
+        assert mock_get_channel_id.call_count == 1
+        # Using get deliberately as there should only be one. Test should fail otherwise.
+        trigger = AlertRuleTrigger.objects.get(alert_rule_id=self.alert_rule.id)
+        action = AlertRuleTriggerAction.objects.get(alert_rule_trigger=trigger)
+        assert action.target_identifier == "10"
+        assert action.target_display == "my-channel"
+
+        # Now two actions with slack:
+        test_params = self.valid_params.copy()
+        test_params["triggers"] = [
+            {
+                "label": "critical",
+                "alertThreshold": 200,
+                "actions": [
+                    {
+                        "type": "slack",
+                        "targetIdentifier": "my-channel",
+                        "targetType": "specific",
+                        "integration": self.integration.id,
+                    },
+                    {
+                        "type": "slack",
+                        "targetIdentifier": "another-channel",
+                        "targetType": "specific",
+                        "integration": self.integration.id,
+                    },
+                    {
+                        "type": "slack",
+                        "targetIdentifier": "another-channel",
+                        "targetType": "specific",
+                        "integration": self.integration.id,
+                    },
+                ],
+            },
+            {
+                "label": "warning",
+                "alertThreshold": 200,
+                "actions": [
+                    {
+                        "type": "slack",
+                        "targetIdentifier": "my-channel",  # same channel, but only one lookup made per channel
+                        "targetType": "specific",
+                        "integration": self.integration.id,
+                    },
+                ],
+            },
+        ]
+
+        with self.feature("organizations:incidents"), self.tasks():
+            resp = self.get_response(
+                self.organization.slug, self.project.slug, self.alert_rule.id, **test_params
+            )
+        assert resp.data["uuid"] == "abc123"
+        assert (
+            mock_get_channel_id.call_count == 3
+        )  # just made 2 calls, plus the call from the single action test
+
+        # Using get deliberately as there should only be one. Test should fail otherwise.
+        triggers = AlertRuleTrigger.objects.filter(alert_rule_id=self.alert_rule.id)
+        actions = AlertRuleTriggerAction.objects.filter(alert_rule_trigger__in=triggers).order_by(
+            "id"
+        )
+        # The 3 critical trigger actions:
+        assert actions[0].target_identifier == "10"
+        assert actions[0].target_display == "my-channel"
+        assert actions[1].target_identifier == "20"
+        assert actions[1].target_display == "another-channel"
+        assert actions[2].target_identifier == "20"
+        assert actions[2].target_display == "another-channel"
+
+        # This is the warning trigger action:
+        assert actions[3].target_identifier == "10"
+        assert actions[3].target_display == "my-channel"
+
+        # Now an invalid action (we want to early out with a good validationerror and not schedule the task):
+        name = "MyInvalidActionRule"
+        test_params["name"] = name
+        test_params["triggers"] = [
+            {
+                "label": "critical",
+                "alertThreshold": 75,
+                "actions": [
+                    {
+                        "type": "element",
+                        "targetIdentifier": "my-channel",
+                        "targetType": "arbitrary",
+                        "integrationId": self.integration.id,
+                    },
+                ],
+            },
+        ]
+        with self.feature("organizations:incidents"), self.tasks():
+            resp = self.get_response(
+                self.organization.slug, self.project.slug, self.alert_rule.id, **test_params
+            )
+        assert resp.status_code == 400
+        assert (
+            mock_get_channel_id.call_count == 3
+        )  # Did not increment from the last assertion because we early out on the validation error
+
 
 @region_silo_test(stable=True)
 class AlertRuleDetailsDeleteEndpointTest(AlertRuleDetailsBase, APITestCase):
     method = "delete"
 
     def test_simple(self):
-        self.create_member(
-            user=self.user, organization=self.organization, role="owner", teams=[self.team]
-        )
-        self.login_as(self.user)
+        self.login_as(self.owner_user)
 
         with self.feature("organizations:incidents"), outbox_runner():
             resp = self.get_success_response(
@@ -550,19 +934,12 @@ class AlertRuleDetailsDeleteEndpointTest(AlertRuleDetailsBase, APITestCase):
         )
 
     def test_no_feature(self):
-        self.create_member(
-            user=self.user, organization=self.organization, role="owner", teams=[self.team]
-        )
-        self.login_as(self.user)
+        self.login_as(self.owner_user)
         self.get_success_response(self.organization.slug, self.alert_rule.id, status_code=204)
 
     def test_snapshot_and_create_new_with_same_name(self):
         with self.tasks():
-            self.create_member(
-                user=self.user, organization=self.organization, role="owner", teams=[self.team]
-            )
-            self.login_as(self.user)
-
+            self.login_as(self.owner_user)
             # We attach the rule to an incident so the rule is snapshotted instead of deleted.
             incident = self.create_incident(alert_rule=self.alert_rule)
